MNT rename _utils._array to _utils.array
