+++ src/model_diagnostics/_utils/array.py
